+++ src/DTO/ResponsiveSource.php
@@ -1,0 +1,34 @@
+<?php
+
+/*
+ * This file is part of the Progressive Image Bundle.
+ *
+ * (c) Jozef Môstka <https://github.com/tito10047/progressive-image-bundle>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Tito10047\ProgressiveImageBundle\DTO;
+
+class ResponsiveSource implements ResponsiveSourceInterface {
+
+	public function __construct(
+		private readonly ?string $media,
+		private readonly string  $srcset,
+		private readonly string  $sizes
+	) {
+	}
+
+	public function getMedia(): ?string {
+		return $this->media;
+	}
+
+	public function getSrcset(): string {
+		return $this->srcset;
+	}
+
+	public function getSizes(): string {
+		return $this->sizes;
+	}
+}